# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Casenova API Documentation",
        "description": "Casenova API Documentation",
        "contact": {
            "email": "admin@casenova.uz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://casenova.lc/api/v1",
            "description": "Api Server"
        }
    ],
    "paths": {
...
        "/galleries": {
            "get": {
                "tags": [
                    "Galleries"
                ],
                "summary": "Get list of galleries",
                "description": "Returns list of galleries",
                "operationId": "getGalleries",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GalleryCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/admin/galleries": {
            "post": {
                "tags": [
                    "Galleries"
                ],
                "summary": "Create new gallery",
                "description": "Returns gallery data",
                "operationId": "storeGalleries",
                "requestBody": {
                    "description": "Pass gallery credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreGalleryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/galleries/{id}": {
            "get": {
                "tags": [
                    "Galleries"
                ],
                "summary": "Get gallery information",
                "description": "Returns gallery data",
                "operationId": "getGalleryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Gallery id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/GalleryResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/admin/galleries/{id}": {
            "put": {
                "tags": [
                    "Galleries"
                ],
                "summary": "Update gallery",
                "description": "Returns gallery data",
                "operationId": "updateGalleries",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Gallery id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass gallery credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGalleryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/admin/galleries/{gallery}": {
            "delete": {
                "tags": [
                    "Galleries"
                ],
                "summary": "Delete gallery",
                "description": "Returns gallery status",
                "operationId": "deleteGallery",
                "parameters": [
                    {
                        "name": "gallery",
                        "in": "path",
                        "description": "Gallery id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/admin/galleries/{gallery}/image": {
            "post": {
                "tags": [
                    "Galleries"
                ],
                "summary": "Update image for gallery",
                "description": "Returns gallery status",
                "operationId": "imageGallery",
                "parameters": [
                    {
                        "name": "gallery",
                        "in": "path",
                        "description": "Gallery id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "$ref": "#/components/schemas/BaseModel/properties/property_uploads"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseModel/properties/property_boolean_result"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreGalleryRequest": {
                "title": "StoreGalleryRequest",
                "properties": {
                    "image": {
                        "$ref": "#/components/schemas/Gallery/properties/image"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreGalleryRequest"
                }
            },
...
            "UpdateGalleryRequest": {
                "title": "UpdateGalleryRequest",
                "properties": {
                    "image": {
                        "$ref": "#/components/schemas/GalleryResource/properties/image"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateGalleryRequest"
                }
            },
...
            "GalleryCollection": {
                "title": "GalleryCollection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GalleryResource"
                        }
                    }
                },
                "type": "object"
            },
            "GalleryResource": {
                "title": "GalleryResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/Gallery/properties/id"
                    },
                    "image": {
                        "$ref": "#/components/schemas/Gallery/properties/image"
                    },
                    "createdAt": {
                        "$ref": "#/components/schemas/Gallery/properties/created_at"
                    },
                    "updatedAt": {
                        "$ref": "#/components/schemas/Gallery/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "GalleryResource"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}